ai-employee-project/pyproject.toml
[project]
name = "ai-employee-bronze"
version = "0.1.0"
description = "Bronze Tier AI Employee - File & Gmail monitoring with agent skills"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "watchdog>=6.0.0",
    "python-frontmatter>=1.1.0",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
]

[project.scripts]
ai-employee = "main:main"
